Speed improved by 10% 	modified:   Makefile 	modified:   lib/cartridge.c 	modified:   lib/cycles.c 	modified:   lib/cycles.h 	modified:   lib/gameboy.c 	modified:   lib/gpu.c 	modified:   lib/gpu.h 	modified:   lib/mmu.c 	modified:   lib/serial.c 	modified:   lib/sound.c 	modified:   lib/sound.h 	modified:   lib/timer.c 	modified:   lib/timer.h 	modified:   lib/z80_gameboy.h 	modified:   pizza.c

+++ lib/gpu.h
@@ -84,17 +84,17 @@ typedef struct gpu_s
     uint8_t   *lyc;
 
     /* clocks counter   */
-    uint32_t  clocks;
+    uint_fast32_t  clocks;
 
     /* gpu step span */
-    uint8_t   step;
+    uint_fast32_t   step;
 
     /* window last drawn lines */
     uint8_t   window_last_ly;
     uint8_t   window_skipped_lines;
 
     /* frame counter */
-    uint16_t  frame_counter;
+    uint_fast16_t  frame_counter;
 
     /* BG palette       */
     uint16_t  bg_palette[4];
